Rename scheduler to kubestellar-scheduler
