+++ cmd/kubestellar-scheduler/cmd/scheduler.go
@@ -36,7 +36,7 @@ import (
 	kcpinformers "github.com/kcp-dev/kcp/pkg/client/informers/externalversions"
 	kcpfeatures "github.com/kcp-dev/kcp/pkg/features"
 
-	scheduleroptions "github.com/kcp-dev/edge-mc/cmd/scheduler/options"
+	scheduleroptions "github.com/kcp-dev/edge-mc/cmd/kubestellar-scheduler/options"
 	edgeclientset "github.com/kcp-dev/edge-mc/pkg/client/clientset/versioned/cluster"
 	edgeinformers "github.com/kcp-dev/edge-mc/pkg/client/informers/externalversions"
 	"github.com/kcp-dev/edge-mc/pkg/scheduler"
