+++ docs/content/en/docs/Coding Milestones/PoC2023q1/example1.md
@@ -357,7 +357,7 @@ deployment.apps/kcp-edge-syncer-florin-32uaph9l created
 "Stage 2 summary")
 
 Stage 2 creates two workloads, called "common" and "special", and lets
-the edge scheduler react.  It has the following steps.
+the scheduler react.  It has the following steps.
 
 ### Create and populate the workload management workspace for the common workload
 
@@ -592,7 +592,7 @@ EOF
 
 ### Edge scheduling
 
-In response to each EdgePlacement, the edge scheduler will create a
+In response to each EdgePlacement, the scheduler will create a
 corresponding SinglePlacementSlice object.  These will indicate the
 following resolutions of the "where" predicates.
 
@@ -601,19 +601,19 @@ following resolutions of the "where" predicates.
 | edge-placement-c | florin, guilder |
 | edge-placement-s | guilder |
 
-Eventually there will be automation that conveniently runs the edge
+Eventually there will be automation that conveniently runs the
 scheduler.  In the meantime, you can run it by hand: switch to the
 ESPW and invoke the edge-mc command that runs the scheduler.
 
 ```console
 $ kubectl ws root:espw
 Current workspace is "root:espw".
-$ scheduler
-I0423 01:33:37.036752   11305 scheduler.go:212] "Found APIExport view" exportName="edge.kcp.io" serverURL="https://192.168.58.123:6443/services/apiexport/7qkse309upzrv0fy/edge.kcp.io"
+$ kubestellar-scheduler
+I0423 01:33:37.036752   11305 kubestellar-scheduler.go:212] "Found APIExport view" exportName="edge.kcp.io" serverURL="https://192.168.58.123:6443/services/apiexport/7qkse309upzrv0fy/edge.kcp.io"
 ...
-I0423 01:33:37.320859   11305 reconcile_on_location.go:192] "updated SinglePlacementSlice" controller="edge-scheduler" triggeringKind=Location key="apmziqj9p9fqlflm|florin" locationWorkspace="apmziqj9p9fqlflm" location="florin" workloadWorkspace="10l175x6ejfjag3e" singlePlacementSlice="edge-placement-c"
+I0423 01:33:37.320859   11305 reconcile_on_location.go:192] "updated SinglePlacementSlice" controller="kubestellar-scheduler" triggeringKind=Location key="apmziqj9p9fqlflm|florin" locationWorkspace="apmziqj9p9fqlflm" location="florin" workloadWorkspace="10l175x6ejfjag3e" singlePlacementSlice="edge-placement-c"
 ...
-I0423 01:33:37.391772   11305 reconcile_on_location.go:192] "updated SinglePlacementSlice" controller="edge-scheduler" triggeringKind=Location key="apmziqj9p9fqlflm|guilder" locationWorkspace="apmziqj9p9fqlflm" location="guilder" workloadWorkspace="10l175x6ejfjag3e" singlePlacementSlice="edge-placement-c"
+I0423 01:33:37.391772   11305 reconcile_on_location.go:192] "updated SinglePlacementSlice" controller="kubestellar-scheduler" triggeringKind=Location key="apmziqj9p9fqlflm|guilder" locationWorkspace="apmziqj9p9fqlflm" location="guilder" workloadWorkspace="10l175x6ejfjag3e" singlePlacementSlice="edge-placement-c"
 ^C
 ```
